+++ android/app/src/main/java/com/mushafcompanion/reader/MainActivity.java
@@ -1,0 +1,5 @@
+package com.mushafcompanion.reader;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}